pyproject.toml
[project]
name = "indgram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.4.1",
    "dart-fss>=0.4.14",
    "foundry-dev-tools>=2.0.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.7",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "openai>=2.14.0",
    "opendartreader>=0.2.3",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
